Fix Typography variant
Apparently "title" is not a valid variant anymore with the current
version of Material UI.

I don't like using h6 here because of accessibility reasons. I would
rather use h1 or no variant and then adjust the size with CSS.
However, as I currently have just a little time for a hotfix, letting
this go through for now. If this was my work project, wouldn't approve
this PR.